requirements.txt
django
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
pillow
six
django-allauth
PyJWT
cryptography
twilio
django-simple-captcha
python-decouple
channels
daphne
